detalle facturas tipos

+++ ArtInk/ArtInk.Application/DTOs/DetalleFacturaProductoDTO.cs
@@ -8,10 +8,10 @@ namespace ArtInk.Application.DTOs;
 
 public record DetalleFacturaProductoDTO
 {
-    public short Id { get; set; }
+    public long Id { get; set; }
 
     [DisplayName("Detalle")]
-    public short IdDetalleFactura { get; set; }
+    public long IdDetalleFactura { get; set; }
 
     [DisplayName("Producto")]
     public short IdProducto { get; set; }
